feat(changelog): extract business logic from app.js
