+++ docs/changelog/index.html
@@ -193,7 +193,7 @@
             <b>Search Examples:</b>
             <ul>
               <li>
-                <a href="?stable_version=1.27.0&package=@webex/widgets">Package search - @webex/widgets@1.27.0</a>
+                <a href="?stable_version=1.28.2&package=@webex/widgets">Package search - @webex/widgets@1.27.0</a>
               </li>
               <li>
                 <a href="?stable_version=1.28.0&package=@webex/cc-widgets&version=1.28.0-ccwidgets.121"
@@ -231,9 +231,9 @@
                 <a href="?compare=1.27.0vs1.28.0">Full version comparison - 1.27.0 vs 1.28.0</a>
               </li>
               <li>
-                <a href="?compareStableA=1.28.0&compareStableB=1.28.2&comparePackage=@webex/widgets&compareVersionA=1.28.2-next.2&compareVersionB=1.28.2-next.3"
-                  >Package-level comparison - @webex/widgets specific versions</a
-                >
+                <a href="?stable_version=1.28.0&package=@webex/cc-widgets&version=1.28.0-ccwidgets.121">
+                  Package + version search - @webex/cc-widgets@1.28.0-ccwidgets.121
+                </a>
               </li>
             </ul>
             <b>How to use version comparison:</b>
@@ -380,6 +380,6 @@
       </div>
     </script>
 
-    <script type="text/javascript" src="assets/js/app.js"></script>
+    <script type="module" src="assets/js/app.js"></script>
   </body>
 </html>
